+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 5.1.0
+
+- Update packages
+- Update success and unsuccessful screens
+
+## 5.0.0
+
+- Updated most of the packages
+- Updated default design
+- Fixed bugs
+- fixed flow
+
+
 ## 4.1.0
 
 - Update to flutter_registration 2.1.0